Extract PlayerScore updates from GameManager into its own script
