+++ Assets/Scripts/UI/GameSession.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82dede14749efc44a94384c263ceee7c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
