+++ Assets/Scripts/UI/GameSession/UpdateScoreText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4c7bea7181cedb4f9fe40394e14e740
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
